% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_data.R
\docType{data}
\name{diagnosis}
\alias{diagnosis}
\title{Diagnosis data}
\format{A data frame with 120 rows and 8 mixed variables.}
\source{
\href{https://archive.ics.uci.edu/ml/datasets/Acute+Inflammations}{UCI Machine Learning, Acute Inflammations Data Set}
}
\usage{
data(diagnosis)
}
\description{
Diagnosis data
}
\keyword{datasets}
