% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlp_para.R
\name{discret}
\alias{discret}
\title{Discretization}
\usage{
discret(colonne, Y)
}
\arguments{
\item{colonne}{the column of a dataframe that will be discretize, must be numeric}

\item{Y}{the column of a dataframe containing modalities to predict}
}
\value{
df_disc a dataframe with the variable colonne discretize, cuts a list giving the cuts made by the mdlp()
function
}
\description{
Discretization using the mdlp() function from the package discretization
}
\examples{
data(iris)
column <- iris[1]
Y <- iris[5]
disc <- discret(column, Y)


}
