% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{discretisation}
\alias{discretisation}
\title{Discretization from cuts}
\usage{
discretisation(column, cuts)
}
\arguments{
\item{column}{the column of a dataframe that will be discretize, must be numeric}

\item{cuts}{a list of cuts made by \emph{mdlp} usually NBAYES$cuts}
}
\value{
a dataframe with all row discretize
}
\description{
Discretization from cuts made by the \emph{mdlp} function from \emph{library(discretization)} during the \emph{fit}
on the training dataset.
}
\examples{
data(iris)
train <- iris
modele <- fit(Species ~ ., train, laplace=0)
disc <- discretisation(iris[1], modele$cuts[[1]])

}
