% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish data}
\format{A data frame with 159 rows and 7 variables}
\source{
\href{https://www.kaggle.com/aungpyaeap/fish-market}{Kaggle, Fish Market}
}
\usage{
data(fish)
}
\description{
Fish data
}
\keyword{datasets}
