% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.NBAYES}
\alias{predict.NBAYES}
\title{Prediction method for NBC}
\usage{
predict.NBAYES(object_NBAYES, data_test, type="both", parallel=FALSE)
}
\arguments{
\item{object_NBAYES}{an object of class NBAYES build using \emph{fit}}

\item{data_test}{a dataframe containing the variables to predict}

\item{type}{a character type, it will define what the function will return. Can be "posterior", if the user want
to get only the posterior probabilities for the different predict class. "class" which give the user only the modality
of the prediction. And "both" that will return both "posterior" and "class" informations.}

\item{parallel}{a logic that allow the user to compute the function using parallel computation.}
}
\value{
a dataframe containing the posterior probabilities, the predict class or both them.
}
\description{
Compute posterior probabilities and class based on Naive Bayes Classifier
}
\examples{
data(iris)
train <- iris
test <- iris[-5]
model <- fit(Species ~ ., train, laplace=1, parallel=FALSE)
res <- predict(model, test, type="class")


}
