% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print.NBAYES}
\alias{print.NBAYES}
\title{Print NBAYES object attributes}
\usage{
print.NBAYES(NBAYES)
}
\arguments{
\item{NBAYES}{an NBAYES object}
}
\value{
print all the different informations about the object NBAYES
}
\description{
print different information about an NBAYES object
}
\examples{
data(iris)
train <- iris
modele <- fit(Species ~ ., train)
print(modele)

}
