% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parall.R
\name{proba_1_obs}
\alias{proba_1_obs}
\title{Posterior probabilities}
\usage{
proba_1_obs(ligne_df, nbayes)
}
\arguments{
\item{ligne_df}{row of the test dataframe to discretize}

\item{nbayes}{the nbayes object return by fit() function}
}
\value{
a vector containing the conditionnal probability for each modality to predict
}
\description{
Get the posterior probabilites from a row of a test dataset using the NBAYES object created while \strong{fit}
\emph{proba_1_obs} gives the probabilities of the observation (row) to belong to the different modalities
of the variable to predict Y.
}
