% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.NBAYES}
\alias{summary.NBAYES}
\title{Summary of a NBAYES object}
\usage{
summary.NBAYES(NBAYES)
}
\arguments{
\item{NBAYES}{an NBAYES object}
}
\value{
print all the different informations about the object NBAYES
}
\description{
summary the information of a NBAYES object
}
\examples{
data(iris)
train <- iris
modele <- fit(Species ~ ., train)
summary(modele)

}
