% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fitting the NBC model}
\usage{
fit(formula, data, laplace=1, parallel=FALSE)
}
\arguments{
\item{formula}{an object of class formula.}

\item{data}{a dataframe containing the variables in the model.}

\item{laplace}{an integer that will be use to do the Laplace smoothing.}

\item{parallel}{a logic that allow the user to compute the function using parallel computation.}
}
\value{
an NBAYES object containing several information that can be useful for the user but more important
for the predict function
}
\description{
fit is used to fit the Naive Bayes Classifier model. It can compute with all types of variables,
applying discretisation on quantitative variables and need a qualitative predictive variable.
}
\examples{
data(iris)
df <- iris
model <- NBC::fit(Species ~ ., data, laplace=1, parallel=FALSE)

}
